+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 
 # Changelog
 
+#### 21th February 2024 - Version 1.8.7
+
+-   Tweak: Added Gallery Banner widget
+
 #### 5th February 2024 - Version 1.8.6
 
 -   Fix: Implement Database updater to update missing widgets due to latest Elementor update (gallery, countdow, blockquote)